+++ CHANGELOG.md
@@ -6,13 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
-
 - Applied the default Syn project meta files ([#70])
 - Rework documentation ([#71])
 - Upgrade to Go 1.15 ([#77])
 - Removed playbook.yml from project and replaced with ad-hoc command ([#79])
 - Update default Steward version ([#82])
 - Make the tenant GitRepo URL required ([#83])
+
+### Added
+- Expose tenant and cluster annotations in API ([#84])
 
 ## [v0.2.0] - 2020-07-23
 ### Changed
@@ -60,3 +62,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#77]: https://github.com/projectsyn/lieutenant-api/pull/77
 [#82]: https://github.com/projectsyn/lieutenant-api/pull/82
 [#83]: https://github.com/projectsyn/lieutenant-api/pull/83
+[#84]: https://github.com/projectsyn/lieutenant-api/pull/84